+++ libs/ui/src/lib/html-section/html-section.component.ts
@@ -1,9 +1,9 @@
 import {
   ChangeDetectionStrategy,
-  Component,
+  Component, ElementRef, HostListener,
   Injector,
   Input,
-  OnInit,
+  OnInit, ViewChild,
 } from '@angular/core';
 import { createCustomElement } from '@angular/elements';
 import { IconComponent } from '../icons/icon/icon.component';
@@ -18,8 +18,18 @@ import { TrackComponent } from '../track/track.component';
 export class HtmlSectionComponent implements OnInit {
   @Input() html?: string;
   @Input() sanitize = true;
+  @ViewChild('htmlSectionRef') htmlSectionRef?: ElementRef;
 
   constructor(private readonly injector: Injector) {}
+
+  @HostListener('document:click') onClick() {
+    if (
+      this.htmlSectionRef &&
+      this.htmlSectionRef.nativeElement.innerHTML !== this.html
+    ) {
+      this.html = this.htmlSectionRef.nativeElement.innerHTML;
+    }
+  }
 
   ngOnInit() {
     this.createCustomElements([