+++ libs/ui/src/lib/tabs/tab-group.component.stories.ts
@@ -3,6 +3,8 @@ import { Meta, moduleMetadata, Story } from '@storybook/angular';
 import { TabGroupComponent } from './tab-group.component';
 import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { UiModule } from '../ui.module';
+import { TabComponent } from './tab.component';
+import { storybookIconsNames } from '../icons/storybook-icons';
 
 export default {
   title: 'Angular/Tabs/Tab group',
@@ -63,5 +65,61 @@ export const Mobile = Template.bind({});
 Mobile.parameters = {
   viewport: {
     defaultViewport: 'iphone12mini',
+  },
+};
+
+const TemplateWithIcon: Story<TabComponent> = (args: TabComponent) => ({
+  props: args,
+  template: `
+<cvi-ng-tab-group>
+  <cvi-ng-tab
+    title="Tab 1"
+    [iconName]="iconName"
+    [svgClass]="'svg-class'"
+    [iconPosition]="iconPosition"
+    [iconHeight]="iconHeight"
+    [gap]="gap"
+    >{{ content }}</cvi-ng-tab
+  >
+  <cvi-ng-tab title="Tab 2 with lots and lots of random text"
+    >Third tab content. File not found. I was having the most wonderful dream.
+    Except you were there, and you were there, and you were there! And when we
+    woke up, we had these bodies. I could if you hadn't turned on the light and
+    shut off my stereo.</cvi-ng-tab
+  >
+</cvi-ng-tab-group>
+  `,
+  styles: [
+    // you don't need to use ::ng-deep in your app
+    `
+      ::ng-deep .svg-class {
+        fill: red;
+      }
+    `,
+  ],
+});
+
+export const WithIcon = TemplateWithIcon.bind({});
+WithIcon.args = {
+  iconName: 'input_error',
+  iconHeight: 15,
+  iconPosition: 'before',
+  gap: 2,
+};
+WithIcon.argTypes = {
+  iconName: {
+    name: 'Icon name',
+    options: storybookIconsNames,
+    control: { type: 'select' },
+  },
+  gap: {
+    name: 'Gap',
+    control: { type: 'range', min: 0, max: 20, step: 1 },
+  },
+  svgClass: { control: false },
+  iconPosition: {
+    name: 'Icon position',
+    options: ['before', 'after'],
+    control: { type: 'select' },
   },
 };